+++ kmonitor_model.go
@@ -1,8 +1,8 @@
 package gokids
 
 import (
-  "sync"
-  "time"
+	"sync"
+	"time"
 
 )
 type KMonitor struct {
@@ -19,6 +19,7 @@ type KMonitorKeyData struct {
 	TotalCount  int64
 	SucessCount int64
 	Duration    int64
+	ProductKey	string		//应用中心产品唯一标志
 }
 
 // 每个请求的处理结果
@@ -26,6 +27,7 @@ type KMonitorReqStatus struct {
 	Key      string // key
 	Status   int    // 处理结果，成功=1，失败=0
 	Duration int64  // 处理时间，ms
+	ProductKey	string		//应用中心产品唯一标志
 }
 type KMonitorReportReqData struct {
 	Code         string `json:"m"`  //服务码
